Перенёс функции для рендеринга в utils/render.js
